Refactor SolKz compositional field benchmark to reduce code duplication
