+++ benchmark/solkz/compositional_field/CMakeLists.txt
@@ -30,10 +30,22 @@ IF (NOT Aspect_FOUND)
         "manual for more information.")
 ENDIF ()
 
+IF (NOT ASPECT_SOURCE_DIR)
+  MESSAGE(FATAL_ERROR "\n"
+    "Please specify the path to the source directory of ASPECT by passing\n"
+	"-DASPECT_SOURCE_DIR=<path to ASPECT source>\n"
+	"to cmake. This is required since solkz_compositional_field is dependent"
+	"on solkz benchmark.")
+ENDIF ()
+
 DEAL_II_INITIALIZE_CACHED_VARIABLES()
 
 SET(TARGET "solkz_compositional_field")
+
 PROJECT(${TARGET})
 
+INCLUDE_DIRECTORIES("${ASPECT_SOURCE_DIR}/benchmark/solkz")
+
 ADD_LIBRARY(${TARGET} SHARED solkz_compositional_field.cc)
-ASPECT_SETUP_PLUGIN(${TARGET})	
+
+ASPECT_SETUP_PLUGIN(${TARGET})
