+++ benchmark/solkz/solkz.cc
@@ -1,16 +1,4 @@
-#include <aspect/material_model/simple.h>
-#include <aspect/velocity_boundary_conditions/interface.h>
-#include <aspect/simulator_access.h>
-#include <aspect/global.h>
-
-#include <deal.II/dofs/dof_tools.h>
-#include <deal.II/numerics/data_out.h>
-#include <deal.II/base/quadrature_lib.h>
-#include <deal.II/base/function_lib.h>
-#include <deal.II/numerics/error_estimator.h>
-#include <deal.II/numerics/vector_tools.h>
-
-
+#include "solkz.h"
 
 namespace aspect
 {
@@ -540,119 +528,20 @@ namespace aspect
           }
       }
 
-
-
-      /**
-       * The exact solution for the SolKz benchmark.
-       */
       template <int dim>
-      class FunctionSolKz : public Function<dim>
+      void FunctionSolKz<dim>::
+      vector_value(const Point<dim> &p, Vector<double> &values) const
       {
-        public:
-          FunctionSolKz () : Function<dim>() {}
-
-          virtual void vector_value (const Point< dim >   &p,
-                                     Vector< double >   &values) const
-          {
-            double pos[2]= {p(0),p(1)};
-            double total_stress[3], strain_rate[3];
-            static const double B = 0.5 * std::log(1e6);
-            AnalyticSolutions::_Velic_solKz
-            (pos,
-             1.0, 2, 3,
-             B,
-             &values[0], &values[2], total_stress, strain_rate );
-          }
-      };
-
-
+        double pos[2] = {p(0), p(1)};
+        double total_stress[3], strain_rate[3];
+        static const double B = 0.5 * std::log(1e6);
+        AnalyticSolutions::_Velic_solKz
+        (pos,
+         1.0, 2, 3,
+         B,
+         &values[0], &values[2], total_stress, strain_rate);
+      }
     }
-
-
-
-    template <int dim>
-    class SolKzMaterial : public MaterialModel::InterfaceCompatibility<dim>
-    {
-      public:
-        /**
-         * @name Physical parameters used in the basic equations
-         * @{
-         */
-        virtual double viscosity (const double                  temperature,
-                                  const double                  pressure,
-                                  const std::vector<double>    &compositional_fields,
-                                  const SymmetricTensor<2,dim> &strain_rate,
-                                  const Point<dim>             &position) const;
-
-        virtual double density (const double temperature,
-                                const double pressure,
-                                const std::vector<double> &compositional_fields,
-                                const Point<dim> &position) const;
-
-        virtual double compressibility (const double temperature,
-                                        const double pressure,
-                                        const std::vector<double> &compositional_fields,
-                                        const Point<dim> &position) const;
-
-        virtual double specific_heat (const double temperature,
-                                      const double pressure,
-                                      const std::vector<double> &compositional_fields,
-                                      const Point<dim> &position) const;
-
-        virtual double thermal_expansion_coefficient (const double      temperature,
-                                                      const double      pressure,
-                                                      const std::vector<double> &compositional_fields,
-                                                      const Point<dim> &position) const;
-
-        virtual double thermal_conductivity (const double temperature,
-                                             const double pressure,
-                                             const std::vector<double> &compositional_fields,
-                                             const Point<dim> &position) const;
-        /**
-         * @}
-         */
-
-        /**
-         * @name Qualitative properties one can ask a material model
-         * @{
-         */
-
-        /**
-         * Return whether the model is compressible or not.
-         * Incompressibility does not necessarily imply that the density is
-         * constant; rather, it may still depend on temperature or pressure.
-         * In the current context, compressibility means whether we should
-         * solve the contuity equation as $\nabla \cdot (\rho \mathbf u)=0$
-         * (compressible Stokes) or as $\nabla \cdot \mathbf{u}=0$
-         * (incompressible Stokes).
-         */
-        virtual bool is_compressible () const;
-        /**
-         * @}
-         */
-
-        /**
-         * @name Reference quantities
-         * @{
-         */
-        virtual double reference_viscosity () const;
-
-        virtual double reference_density () const;
-
-        virtual double reference_thermal_expansion_coefficient () const;
-
-//TODO: should we make this a virtual function as well? where is it used?
-        double reference_thermal_diffusivity () const;
-
-        double reference_cp () const;
-        /**
-         * @}
-         */
-
-        void
-        parse_parameters (ParameterHandler &prm);
-    };
-
 
 
     template <int dim>
@@ -792,28 +681,6 @@ namespace aspect
     }
 
 
-
-
-
-
-    /**
-      * A postprocessor that evaluates the accuracy of the solution.
-      *
-      * The implementation of error evaluators that correspond to the
-      * benchmarks defined in the paper Duretz et al. reference above.
-      */
-    template <int dim>
-    class SolKzPostprocessor : public Postprocess::Interface<dim>, public ::aspect::SimulatorAccess<dim>
-    {
-      public:
-        /**
-         * Generate graphical output from the current solution.
-         */
-        virtual
-        std::pair<std::string,std::string>
-        execute (TableHandler &statistics);
-    };
-
     template <int dim>
     std::pair<std::string,std::string>
     SolKzPostprocessor<dim>::execute (TableHandler &statistics)
@@ -841,8 +708,9 @@ namespace aspect
       Vector<float> cellwise_errors_pl2 (this->get_triangulation().n_active_cells());
 
       ComponentSelectFunction<dim> comp_u(std::pair<unsigned int, unsigned int>(0,dim),
-                                          dim+2);
-      ComponentSelectFunction<dim> comp_p(dim, dim+2);
+                                          this->get_fe().n_components());
+      ComponentSelectFunction<dim> comp_p(dim,
+                                          this->get_fe().n_components());
 
       VectorTools::integrate_difference (this->get_mapping(),this->get_dof_handler(),
                                          this->get_solution(),
@@ -886,7 +754,6 @@ namespace aspect
 
       return std::make_pair("Errors u_L1, p_L1, u_L2, p_L2:", os.str());
     }
-
   }
 }
